+++ package/components/inline-comment/use-comment-draft-auto-submit-countdown.ts
@@ -1,6 +1,5 @@
-import { useCallback, useEffect, useRef, useState } from 'react';
-
-const AUTO_SUBMIT_COUNTDOWN_SECONDS = 5;
+import { useCallback, useEffect, useState } from 'react';
+import { useAutoSubmitCountdown } from './use-auto-submit-countdown';
 
 interface UseCommentDraftAutoSubmitCountdownProps {
   draftId: string | null;
@@ -14,12 +13,6 @@ export const useCommentDraftAutoSubmitCountdown = ({
   onSubmit,
 }: UseCommentDraftAutoSubmitCountdownProps) => {
   const [isDraftFocused, setIsDraftFocused] = useState(true);
-  const [remainingSeconds, setRemainingSeconds] = useState(
-    AUTO_SUBMIT_COUNTDOWN_SECONDS,
-  );
-  const hasSubmittedRef = useRef(false);
-
-  const shouldRunCountdown = canAutoSubmit && !isDraftFocused;
 
   const handleDraftFocus = useCallback(() => {
     setIsDraftFocused(true);
@@ -30,42 +23,19 @@ export const useCommentDraftAutoSubmitCountdown = ({
   }, []);
 
   useEffect(() => {
-    hasSubmittedRef.current = false;
     setIsDraftFocused(true);
-    setRemainingSeconds(AUTO_SUBMIT_COUNTDOWN_SECONDS);
   }, [draftId]);
 
-  useEffect(() => {
-    if (!shouldRunCountdown) {
-      hasSubmittedRef.current = false;
-      setRemainingSeconds((current) =>
-        current === AUTO_SUBMIT_COUNTDOWN_SECONDS
-          ? current
-          : AUTO_SUBMIT_COUNTDOWN_SECONDS,
-      );
-      return;
-    }
-
-    const timeoutId = window.setTimeout(() => {
-      if (remainingSeconds <= 1) {
-        if (!hasSubmittedRef.current) {
-          hasSubmittedRef.current = true;
-          onSubmit();
-        }
-        return;
-      }
-
-      setRemainingSeconds(remainingSeconds - 1);
-    }, 1000);
-
-    return () => {
-      window.clearTimeout(timeoutId);
-    };
-  }, [onSubmit, remainingSeconds, shouldRunCountdown]);
+  const { submitLabel } = useAutoSubmitCountdown({
+    label: 'Send',
+    onSubmit,
+    resetKey: draftId,
+    shouldRun: canAutoSubmit && !isDraftFocused,
+  });
 
   return {
     handleDraftBlur,
     handleDraftFocus,
-    submitLabel: `Send ( ${remainingSeconds} )`,
+    submitLabel,
   };
 };
